requirements.txt
fastmcp>=2.0.0
starlette>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.0
uvicorn>=0.27.0
pydantic>=2.0.0

